Tentative conflict resolution
