+++ App/frontend/package.json
