+++ App/frontend/public/manifest.json
