+++ App/frontend/tailwind.config.js
